Added some new features and solved some bugs

+++ API/Controller/pondController.js
@@ -1,16 +1,18 @@
 const Ponds = require('../models/Ponds');
+const Farms = require('../models/Farms');
+const farmsFunctions = require('../functions/farms');
+
 
 exports.newPond = async(req,res,next)=>{
 
 const Pond = new Ponds(req.body);
-
-try {
-    await Pond.save();
-    res.json({message:'New Pond saved succesfully'});
-} catch (error) {
-    console.log(error);
-    next();
-}
+    try {
+        await Pond.save();
+        res.json({message:'New Pond saved succesfully'});
+    } catch (error) {
+        console.log(error);
+        next();
+    }
 };
 
 exports.showPonds =async(req,res,next)=>{
@@ -26,17 +28,17 @@ exports.showPonds =async(req,res,next)=>{
 
 exports.showPond = async(req,res,next)=>{
 
-try {
-    const pond =await Ponds.findById(req.params.idPond);
-    if (!pond) {
-        res.json({message: 'Pond did not founded.'});
+    try {
+        const pond =await Ponds.findById(req.params.idPond);
+        if (!pond) {
+            res.json({message: 'Pond did not founded.'});
+            next();
+        }
+        res.json(pond);
+    } catch (error) {
+        console.log(error);
         next();
     }
-    res.json(pond);
-} catch (error) {
-    console.log(error);
-    next();
-}
 }
 
 exports.updatePond =async(req,res,next)=>{
@@ -56,10 +58,15 @@ exports.updatePond =async(req,res,next)=>{
 
 exports.deletePond= async(req,res,next)=>{
     try {
+      if (  !farmsFunctions.UpdateinDeletePond(req.params.idPond)) {
+        res.json({message: 'Pond did not founded.'});
+        next();
+      }
         await Ponds.findOneAndDelete(
-            {_id:req.params.idPond}
-        );
-        res.json({message:"Pond succesfuly deleted!"})
+                {_id:req.params.idPond}
+            );
+              
+            res.json({message:"Pond succesfuly deleted!"})
     } catch (error) {
         console.log(error);
         next();
@@ -78,3 +85,4 @@ exports.findPond= async(req,res,next)=>{
         next();
     }
 }
+
